Move planning scenarios handler into blueprint
