+++ tests/test_operations_blueprint.py
@@ -35,3 +35,10 @@ def test_coverage_handler_is_no_longer_in_legacy_module():
 
     source = (Path(__file__).parents[1] / "app.py").read_text()
     assert "def coverage_heatmap(" not in source
+
+
+def test_scenarios_handler_is_no_longer_in_legacy_module():
+    from pathlib import Path
+
+    source = (Path(__file__).parents[1] / "app.py").read_text()
+    assert "def scenarios_page():" not in source
